Ajax chart and timezone flexibility
The chart will be now get the information from an ajax request and
it will be drawn when the information is retrieved from zendesk api.

Added a bugfix related to timezone flexibility where we were looking
for ticket updates in the current UTC day and producing wrong results.
The app will now retrieve all time logged in user-assigned tickets
during the last 12 hours.